Made branch Lars and made some calculations in endweek for harvesting, phosphorizing fields and feeding citizens in city.

+++ woz/business/implementations/PlayerImplementation.java
@@ -3,16 +3,16 @@ package business.implementations;
 import business.interfaces.Player;
 
 public class PlayerImplementation implements Player {
-    public int money;
+    public double money;
     @Override
     public void spawn() {
-        this.money = 150;
+        this.money = 150.00;
     }
 
     @Override
-    public void addMoney(int money) {
+    public void addMoney(double money) {
         this.money += money;
     }
     @Override
-    public void useMoney(int money) { this.money -= money; }
+    public void useMoney(double money) { this.money -= money; }
 }
