+++ woz/business/interfaces/Player.java
@@ -3,6 +3,7 @@ package business.interfaces;
 public interface Player {
     void spawn();
 
-    void addMoney(int money);
-    void useMoney(int money);
+    void addMoney(double money);
+    void useMoney(double money);
+
 }
